Refactoring material handling
Made the materials handled by a separate "database" module to bring the responsibility out of the recipe list module.

Seemed to fix a lot of weird quirks from before. Will need to do the same for the operator selection so each part of them can be highlighted when there's enough stuff